cleans up consensus
