+++ crates/consensus/src/transaction/mod.rs
@@ -1,3 +1,5 @@
+//! Tramsaction types for Optimism.
+
 mod deposit;
 pub use deposit::TxDeposit;
 
@@ -17,4 +19,33 @@ pub use source::{
 #[cfg(all(feature = "serde", feature = "serde-bincode-compat"))]
 pub(super) mod serde_bincode_compat {
     pub use super::deposit::serde_bincode_compat::TxDeposit;
+}
+
+use alloy_primitives::B256;
+
+/// A trait representing a deposit transaction with specific attributes.
+pub trait DepositTransaction {
+    /// Returns the hash that uniquely identifies the source of the deposit.
+    ///
+    /// # Returns
+    /// An `Option<B256>` containing the source hash if available.
+    fn source_hash(&self) -> Option<B256>;
+
+    /// Returns the optional mint value of the deposit transaction.
+    ///
+    /// # Returns
+    /// An `Option<u128>` representing the ETH value to mint on L2, if any.
+    fn mint(&self) -> Option<u128>;
+
+    /// Indicates whether the transaction is exempt from the L2 gas limit.
+    ///
+    /// # Returns
+    /// A `bool` indicating if the transaction is a system transaction.
+    fn is_system_transaction(&self) -> bool;
+
+    /// Checks if the transaction is a deposit transaction.
+    ///
+    /// # Returns
+    /// A `bool` that is always `true` for deposit transactions.
+    fn is_deposit(&self) -> bool;
 }
